register more scopes in session
